requirements.txt
openai>=1.44.0
anthropic>=0.25.5
pyaudio
python-dotenv
sounddevice
numpy